Убрано правило headers only
