+++ .vs/VSWorkspaceState.json
@@ -2,6 +2,6 @@
   "ExpandedNodes": [
     ""
   ],
-  "SelectedNode": "\\AiThinker.h",
+  "SelectedNode": "\\ugolki2.sln",
   "PreviewInSolutionExplorer": false
 }
