--- allegro_headers.h
@@ -1,4 +1,0 @@
-#include "allegro5/allegro.h"
-#include "allegro5/allegro_ttf.h"
-#include "allegro5/allegro_primitives.h"
-#include "allegro5/allegro_image.h"
