+++ x64/Release/ugolki2.Build.CppClean.log
@@ -1,0 +1,25 @@
+c:\users\denfe\source\repos\corners_test\x64\release\vc142.pdb
+c:\users\denfe\source\repos\corners_test\x64\release\game.obj
+c:\users\denfe\source\repos\corners_test\x64\release\sysmain.obj
+c:\users\denfe\source\repos\corners_test\x64\release\gamelogicdata.obj
+c:\users\denfe\source\repos\corners_test\x64\release\gamelogicbase.obj
+c:\users\denfe\source\repos\corners_test\x64\release\gamelogibase.obj
+c:\users\denfe\source\repos\corners_test\x64\release\basegamegs.obj
+c:\users\denfe\source\repos\corners_test\x64\release\aiturngs.obj
+c:\users\denfe\source\repos\corners_test\x64\release\aithinker.obj
+c:\users\denfe\source\repos\corners_test\x64\release\mainscreengs.obj
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.exe
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.pdb
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.res
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.ipdb
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.iobj
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\cl.command.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\cl.read.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\cl.write.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\link.command.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\link.read.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\link.write.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\rc.command.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\rc.read.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\rc.write.1.tlog
+c:\users\denfe\source\repos\corners_test\x64\release\ugolki2.tlog\ugolki2.write.1u.tlog
